Move calculateAvgGradeOfClassworkAssignmentInCourse function
