+++ packages/server/schema.gql
@@ -68,6 +68,88 @@ type Course implements BaseModel {
   createdByAccountId: ID!
 }
 
+type ClassworkAssignment implements BaseModel {
+  id: ID!
+  orgId: ID!
+  createdAt: DateTime!
+  updatedAt: DateTime!
+  createdByAccountId: String!
+  courseId: ID!
+  title: String!
+  type: String!
+  description: String
+  attachments: [String!]!
+  publicationState: String!
+  resolveType: [String!]!
+  comments(commentPageOptionInput: CommentPageOptionInput!, lastId: ID): CommentsPayload!
+  dueDate: DateTime
+  maxScores: Float!
+}
+
+input CommentPageOptionInput {
+  limit: Int!
+}
+
+type ClassworkMaterial implements BaseModel {
+  id: ID!
+  orgId: ID!
+  createdAt: DateTime!
+  updatedAt: DateTime!
+  createdByAccountId: String!
+  courseId: ID!
+  title: String!
+  type: String!
+  description: String
+  attachments: [String!]!
+  publicationState: String!
+  resolveType: [String!]!
+  comments(commentPageOptionInput: CommentPageOptionInput!, lastId: ID): CommentsPayload!
+}
+
+type ClassworkAssignmentPayload {
+  classworkAssignments: [ClassworkAssignment!]!
+  count: Int!
+}
+
+type ClassworkMaterialPayload {
+  classworkMaterials: [ClassworkMaterial!]!
+  count: Int!
+}
+
+type AvgGradeOfClassworkByCourse {
+  classworkTitle: String!
+  avgGrade: Float!
+}
+
+type File implements BaseModel {
+  id: ID!
+  orgId: ID!
+  createdAt: DateTime!
+  updatedAt: DateTime!
+  name: String!
+  size: Int!
+  mimeType: String!
+  storageProvider: FileLocation!
+  storageProviderIdentifier: String!
+  signedUrl: String
+}
+
+enum FileLocation {
+  LocalStorage
+}
+
+type ClassworkSubmission implements BaseModel {
+  id: ID!
+  orgId: ID!
+  createdAt: DateTime!
+  updatedAt: DateTime!
+  createdByAccountId: String!
+  classworkId: ID!
+  grade: Float!
+  submissionFileIds: [String!]!
+  description: String!
+}
+
 type Account implements BaseModel {
   id: ID!
   orgId: ID!
@@ -157,23 +239,6 @@ type OrgAccountsPayload {
   count: Int!
 }
 
-type File implements BaseModel {
-  id: ID!
-  orgId: ID!
-  createdAt: DateTime!
-  updatedAt: DateTime!
-  name: String!
-  size: Int!
-  mimeType: String!
-  storageProvider: FileLocation!
-  storageProviderIdentifier: String!
-  signedUrl: String
-}
-
-enum FileLocation {
-  LocalStorage
-}
-
 type AcademicSubjectsPayload {
   academicSubjects: [AcademicSubject!]!
   count: Int!
@@ -182,56 +247,6 @@ type AcademicSubjectsPayload {
 type CoursesPayload {
   courses: [Course!]!
   count: Int!
-}
-
-type ClassworkAssignment implements BaseModel {
-  id: ID!
-  orgId: ID!
-  createdAt: DateTime!
-  updatedAt: DateTime!
-  createdByAccountId: String!
-  courseId: ID!
-  title: String!
-  type: String!
-  description: String
-  attachments: [String!]!
-  publicationState: String!
-  resolveType: [String!]!
-  comments(commentPageOptionInput: CommentPageOptionInput!, lastId: ID): CommentsPayload!
-  dueDate: DateTime
-  maxScores: Float!
-}
-
-input CommentPageOptionInput {
-  limit: Int!
-}
-
-type ClassworkMaterial implements BaseModel {
-  id: ID!
-  orgId: ID!
-  createdAt: DateTime!
-  updatedAt: DateTime!
-  createdByAccountId: String!
-  courseId: ID!
-  title: String!
-  type: String!
-  description: String
-  attachments: [String!]!
-  publicationState: String!
-  resolveType: [String!]!
-  comments(commentPageOptionInput: CommentPageOptionInput!, lastId: ID): CommentsPayload!
-}
-
-type ClassworkSubmission implements BaseModel {
-  id: ID!
-  orgId: ID!
-  createdAt: DateTime!
-  updatedAt: DateTime!
-  createdByAccountId: String!
-  classworkId: ID!
-  grade: Float!
-  submissionFileIds: [String!]!
-  description: String!
 }
 
 type Comment implements BaseModel {
@@ -247,21 +262,6 @@ type Comment implements BaseModel {
 type CommentsPayload {
   comments: [Comment!]!
   count: Int!
-}
-
-type ClassworkAssignmentPayload {
-  classworkAssignments: [ClassworkAssignment!]!
-  count: Int!
-}
-
-type ClassworkMaterialPayload {
-  classworkMaterials: [ClassworkMaterial!]!
-  count: Int!
-}
-
-type AvgGradeOfClassworkByCourse {
-  classworkTitle: String!
-  avgGrade: Float!
 }
 
 type Query {
@@ -327,6 +327,7 @@ type Mutation {
   addStudentsToCourse(studentIds: [ID!]!, courseId: ID!): Course!
   removeStudentsFromCourse(studentIds: [ID!]!, id: ID!): Course!
   removeLecturersFromCourse(lecturerIds: [ID!]!, id: ID!): Course!
+  calculateAvgGradeOfClassworkAssignmentInCourse(optionInput: AvgGradeOfClassworkByCourseOptionInput!, courseId: ID!): [AvgGradeOfClassworkByCourse!]!
   createOrgOffice(input: CreateOrgOfficeInput!): OrgOffice!
   updateOrgOffice(input: UpdateOrgOfficeInput!, id: ID!): OrgOffice!
   findOrgOffices(searchText: String, orgId: ID): [OrgOffice!]!
@@ -341,8 +342,7 @@ type Mutation {
   updateClassworkAssignmentPublication(publication: String!, id: ID!): ClassworkAssignment!
   addAttachmentsToClassworkAssignment(attachmentsInput: AddAttachmentsToClassworkInput!, classworkAssignmentId: ID!): ClassworkAssignment!
   removeAttachmentsFromClassworkAssignments(attachments: [String!]!, classworkAssignmentId: ID!): ClassworkAssignment!
-  calculateAvgGradeOfClassworkAssignmentInCourse(optionInput: AvgGradeOfClassworkByCourseOptionInput!, courseId: ID!): [AvgGradeOfClassworkByCourse!]!
-  createClassworkSubmission(CreateClassworkSubmissionInput: CreateClassworkSubmissionInput!, courseId: ID!): ClassworkSubmission!
+  createClassworkSubmission(createClassworkSubmissionInput: CreateClassworkSubmissionInput!, courseId: ID!): ClassworkSubmission!
   setGradeForClassworkSubmission(setGradeForClassworkSubmissionInput: SetGradeForClassworkSubmissionInput!, courseId: ID!): ClassworkSubmission!
   createComment(commentInput: CreateCommentInput!): Comment!
 }
@@ -394,6 +394,10 @@ input UpdateCourseInput {
   lecturerIds: [ID!]
 }
 
+input AvgGradeOfClassworkByCourseOptionInput {
+  limit: Float = 0
+}
+
 input CreateOrgOfficeInput {
   name: String!
   address: String!
@@ -434,10 +438,6 @@ input UpdateClassworkAssignmentInput {
   title: String
   description: String
   dueDate: String
-}
-
-input AvgGradeOfClassworkByCourseOptionInput {
-  limit: Float = 0
 }
 
 input CreateClassworkSubmissionInput {
