+++ packages/server/src/modules/academic/academic.service.spec.ts
@@ -4,6 +4,8 @@ import { Connection } from 'mongoose'
 import { Publication } from 'core'
 import { objectId } from 'core/utils/db'
 import { createTestingModule, initTestDb } from 'core/utils/testing'
+import { ClassworkService } from 'modules/classwork/classwork.service'
+import { AvgGradeOfClassworkByCourseOptionInput } from 'modules/classwork/classwork.type'
 import { OrgOfficeService } from 'modules/orgOffice/orgOffice.service'
 import { ANY } from 'types'
 
@@ -22,6 +24,7 @@ describe('academic.service', () => {
   let authService: AuthService
   let accountService: AccountService
   let mongooseConnection: Connection
+  let classworkService: ClassworkService
 
   beforeAll(async () => {
     const testDb = await initTestDb()
@@ -34,6 +37,7 @@ describe('academic.service', () => {
     orgOfficeService = module.get<OrgOfficeService>(OrgOfficeService)
     authService = module.get<AuthService>(AuthService)
     accountService = module.get<AccountService>(AccountService)
+    classworkService = module.get<ClassworkService>(ClassworkService)
   })
 
   afterAll(async () => {
@@ -625,6 +629,7 @@ describe('academic.service', () => {
       })
     })
   })
+
   describe('Course', () => {
     const createCourseInput: ANY = {
       academicSubjectId: objectId(),
@@ -2166,4 +2171,204 @@ describe('academic.service', () => {
       })
     })
   })
+
+  /**
+   * START STATISTICAL
+   */
+
+  describe('Statistical', () => {
+    describe('calculateAvgGradeOfClassworkAssignmentInCourse', () => {
+      const optionInput: AvgGradeOfClassworkByCourseOptionInput = {
+        limit: 0,
+      }
+
+      it('throws error if OrgId invalid', async () => {
+        expect.assertions(1)
+
+        jest
+          .spyOn(academicService['orgService'], 'validateOrgId')
+          .mockResolvedValueOnce(false as ANY)
+
+        await expect(
+          academicService.calculateAvgGradeOfClassworkAssignmentInCourse(
+            objectId(),
+            objectId(),
+            optionInput,
+          ),
+        ).rejects.toThrowError('ORG_ID_INVALID')
+      })
+
+      it('throws error if course not found', async () => {
+        expect.assertions(1)
+
+        jest
+          .spyOn(academicService['orgService'], 'validateOrgId')
+          .mockResolvedValueOnce(true as ANY)
+
+        await expect(
+          academicService.calculateAvgGradeOfClassworkAssignmentInCourse(
+            objectId(),
+            objectId(),
+            optionInput,
+          ),
+        ).rejects.toThrowError('COURSE_NOT_FOUND')
+      })
+
+      it('returns an array of data', async () => {
+        expect.assertions(1)
+
+        const org = await orgService.createOrg({
+          namespace: 'kmin-edu',
+          name: 'Kmin Academy',
+        })
+
+        const accStudent = await accountService.createAccount({
+          orgId: org.id,
+          email: 'huynhthanhcanh.top@gmail.com',
+          password: '123456',
+          username: 'thanhthanh',
+          roles: ['student'],
+          displayName: 'Huynh Thanh Thanh',
+        })
+
+        const accStudent2 = await accountService.createAccount({
+          orgId: org.id,
+          email: 'huynhthanhcanh22.top@gmail.com',
+          password: '123456',
+          username: 'thanhthanh2',
+          roles: ['student'],
+          displayName: 'Huynh Thanh Thanh',
+        })
+
+        const accLecturer = await accountService.createAccount({
+          roles: ['lecturer'],
+          email: 'huynhthanhcanh1.top@gmail.com',
+          username: 'thanhcanh',
+          orgId: org.id,
+          displayName: 'Huynh Thanh Canh',
+        })
+
+        const accAdmin = await accountService.createAccount({
+          roles: ['admin'],
+          username: 'thanhcanh123',
+          orgId: org.id,
+          email: 'huynhthanhcanh3.top@gmail.com',
+          displayName: 'Huynh Thanh Canh Thanh',
+        })
+
+        const academicSubject = await academicService.createAcademicSubject({
+          orgId: org.id,
+          code: 'NODEJS',
+          name: 'NodeJS',
+          description: 'This is NodeJs',
+          createdByAccountId: accAdmin.id,
+          imageFileId: objectId(),
+        })
+
+        const orgOffice = await orgOfficeService.createOrgOffice({
+          name: 'Kmin Quận 1',
+          address: '25A',
+          createdByAccountId: accAdmin.id,
+          orgId: org.id,
+          phone: '0704917152',
+        })
+
+        const createCourseInput: ANY = {
+          academicSubjectId: academicSubject.id,
+          orgOfficeId: orgOffice.id,
+          code: 'NodeJS-12',
+          name: 'Node Js Thang 12',
+          tuitionFee: 5000000,
+          lecturerIds: [accLecturer.id],
+        }
+
+        const course = await academicService.createCourse(accAdmin.id, org.id, {
+          ...createCourseInput,
+          startDate: Date.now(),
+          lecturerIds: [accLecturer.id],
+        })
+
+        course.studentIds = [accStudent.id, accStudent2.id]
+        course.save()
+
+        const classwork1 = await classworkService.createClassworkAssignment(
+          accLecturer.id,
+          course.id,
+          org.id,
+          {
+            title: 'Bai tap 1',
+            description: 'Day la bai tap 1',
+          },
+        )
+
+        await classworkService.createClassworkAssignment(
+          accLecturer.id,
+          course.id,
+          org.id,
+          {
+            title: 'Bai tap 2',
+            description: 'Day la bai tap 2',
+          },
+        )
+
+        await classworkService.createClassworkAssignment(
+          accLecturer.id,
+          course.id,
+          org.id,
+          {
+            title: 'Bai tap 3',
+            description: 'Day la bai tap 3',
+          },
+        )
+
+        await authService.signIn({
+          password: 'huynhthanhcanh1.top@gmail.com',
+          usernameOrEmail: 'huynhthanhcanh1.top@gmail.com',
+          orgNamespace: 'kmin-edu',
+        })
+
+        const createInputWithFile: ANY = {
+          classworkId: classwork1.id,
+        }
+
+        const classWorkSubmission1 =
+          await classworkService.createClassworkSubmission(
+            org.id,
+            course.id,
+            accStudent.id,
+            createInputWithFile,
+          )
+
+        classWorkSubmission1.grade = 60
+        classWorkSubmission1.save()
+
+        const classWorkSubmission2 =
+          await classworkService.createClassworkSubmission(
+            org.id,
+            course.id,
+            accStudent.id,
+            createInputWithFile,
+          )
+
+        classWorkSubmission2.grade = 50
+        classWorkSubmission2.save()
+
+        await expect(
+          academicService.calculateAvgGradeOfClassworkAssignmentInCourse(
+            course.id,
+            org.id,
+            optionInput,
+          ),
+        ).resolves.toMatchObject([
+          { avgGrade: 0, classworkTitle: 'Bai tap 3' },
+          { avgGrade: 0, classworkTitle: 'Bai tap 2' },
+          { avgGrade: 55, classworkTitle: 'Bai tap 1' },
+        ])
+      })
+    })
+  })
+
+  /**
+   * END STATISTICAL
+   */
 })
