+++ packages/server/src/modules/classwork/classwork.service.ts
@@ -1032,70 +1032,4 @@ export class ClassworkService {
   /**
    * END CLASSWORK SUBMISSION
    */
-
-  /**
-   * START STATISTICAL
-   */
-
-  async calculateAvgGradeOfClassworkAssignmentInCourse(
-    courseId: string,
-    orgId: string,
-    optionInput: AvgGradeOfClassworkByCourseOptionInput,
-  ): Promise<AvgGradeOfClassworkByCourse[]> {
-    const { limit } = optionInput
-
-    const { courseModel, classworkAssignmentsModel } = this
-
-    if (!(await this.orgService.validateOrgId(orgId))) {
-      throw new Error(`ORG_ID_INVALID`)
-    }
-
-    const course = await courseModel.findOne({
-      _id: courseId,
-      orgId,
-    })
-
-    if (!course) {
-      throw new Error(`COURSE_NOT_FOUND`)
-    }
-
-    const numberOfStudent = course.studentIds.length
-
-    const classworkAssignments = await classworkAssignmentsModel
-      .find({
-        courseId,
-        orgId,
-      })
-      .sort({ _id: -1 })
-      .limit(limit)
-
-    const dataList: AvgGradeOfClassworkByCourse[] = []
-
-    const classworkAssignmentsMap = classworkAssignments.map(
-      async (classworkAssignment) => {
-        const avgGrade = await this.calculateAvgGradeOfClassworkAssignment(
-          numberOfStudent,
-          classworkAssignment.id,
-          orgId,
-        )
-        const dataObj: AvgGradeOfClassworkByCourse = {
-          classworkTitle: classworkAssignment.title,
-          avgGrade,
-        }
-        return dataObj
-      },
-    )
-
-    await Promise.all(classworkAssignmentsMap).then((dataObj) => {
-      dataObj.forEach((data) => {
-        dataList.push(data)
-      })
-    })
-
-    return dataList
-  }
-
-  /**
-   * END STATISTICAL
-   */
 }
